Fixed a problem with overtick during interruptsa
If an overtick was generated when in kernel mode, during an interrupt,
the flow control was such that the value of rip was pushed twice (or
more) on stack. We now check in a safer way whether we are running the
hook during an interrupt handler. In this way, we might loose some
overticks, but atomicity of the handler is ensured.