Fix a race condition in the kernel module
In case an overtick is received while in kernel mode, upon executing
an iret the control was given again to the overtick module. In some
circumstances, this would have placed twice a return address on stack,
while the application-level handler would have been called only once.
In this case, the stack would have been misaligned.

Signed-off-by: Francesco Quaglia <quaglia@dis.uniroma1.it>