Pass the first byte off .text when registering
The library has access to __data_start. It's address is passed to
the kernel module when registering a thread. In this way, we can
perform a fast check to determine whether a certain IP value is
related to userspace program, or the kernel/libraries.

There could still be some problems with got/plt.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>